requirements.txt
addict==2.4.0
attr==0.3.2
backcall==0.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.4.0
cycler==0.11.0
decorator==5.1.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.2
kiwisolver==1.4.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
MedPy==0.4.0
mmcv-full==1.3.0
mmsegmentation==0.12.0
numpy==1.21.6
opencv-python==4.10.0.84
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.5
scipy==1.7.3
SimpleITK==2.3.0
six==1.16.0
terminaltables==3.1.10
timm==0.9.12
tomli==2.0.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.67.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
yapf==0.40.1
zipp==3.15.0
